Refactor code to regenerate newsletter content
This commit refactors the code to add the functionality to regenerate the content of a newsletter. It introduces a new API endpoint `/api/regenerate-newsletter` that takes the `newsletterId` and `userId` as parameters. The code fetches the existing newsletter and its associated used articles. It then regenerates the content using the `generatePersonalizedContent` function, passing in the set of used articles. The content is updated in the database, along with the content history and the regenerate count. The first four articles are also checked for duplication and added to the used articles if they are not already present.