Refactor Email Query Hook
Separated the concerns of fetching emails and newsletters into two distinct hooks. The useNewsletterQuery hook was previously responsible for fetching both emails and newsletters. This commit splits these responsibilities, creating a new useEmailQuery hook for fetching emails. This change improves the modularity and readability of the code, making it easier to maintain and understand.